backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gitpython==3.1.43
langchain-text-splitters==0.3.0
pydantic==2.9.2
aiosqlite==0.20.0
python-multipart==0.0.12
requests==2.32.3
faiss-cpu==1.9.0
sentence-transformers==3.3.1
